Delete old ExtensionPoint file and rename ExtensionPointFC. Move isSiteEditorIframe to utils/dom. Move site editor logic from ExtensionPoint to ExtensionPointComponent.
