+++ react/components/ExtensionPointComponent.tsx
@@ -1,3 +1,4 @@
+import ReactDOM from 'react-dom'
 import * as Sentry from '@sentry/browser'
 import PropTypes from 'prop-types'
 import { forEachObjIndexed, pickBy } from 'ramda'
@@ -10,6 +11,7 @@ import ExtensionPointError from './ExtensionPointError'
 import Loading from './Loading'
 import { RenderContextProps } from './RenderContext'
 import { TreePathContext } from '../utils/treePath'
+import { isSiteEditorIframe } from '../utils/dom'
 
 interface Props {
   component: string | null
@@ -160,6 +162,7 @@ class ExtensionPointComponent extends PureComponent<
   public componentDidMount() {
     this._isMounted = true
     this.fetchAndRerender()
+    this.addDataToElementIfEditable()
   }
 
   public componentDidUpdate() {
@@ -171,9 +174,11 @@ class ExtensionPointComponent extends PureComponent<
         this.mountedError = true
       }
     }
+    this.addDataToElementIfEditable()
   }
 
   public componentWillUnmount() {
+    this.removeDataFromElement()
     this._isMounted = false
   }
 
@@ -223,6 +228,40 @@ class ExtensionPointComponent extends PureComponent<
         )}
       </TreePathContext.Provider>
     )
+  }
+
+  private addDataToElementIfEditable = () => {
+    if (!isSiteEditorIframe) {
+      return
+    }
+    const ComponentImpl = this.props.component && getImplementation(this.props.component)
+    const isEditable =
+      ComponentImpl &&
+      (ComponentImpl.hasOwnProperty('schema') ||
+        ComponentImpl.hasOwnProperty('getSchema'))
+
+    if (!isEditable) {
+      return
+    }
+
+    // eslint-disable-next-line react/no-find-dom-node
+    const element = ReactDOM.findDOMNode(this) as Element
+
+    if (element && element.setAttribute) {
+      element.setAttribute('data-extension-point', this.props.treePath)
+    }
+  }
+
+  private removeDataFromElement = () => {
+    if (!isSiteEditorIframe) {
+      return
+    }
+    // eslint-disable-next-line react/no-find-dom-node
+    const element = ReactDOM.findDOMNode(this) as Element
+
+    if (element && element.removeAttribute) {
+      element.removeAttribute('data-extension-point')
+    }
   }
 }
