+++ react/components/LayoutContainer.tsx
@@ -1,8 +1,7 @@
 import PropTypes from 'prop-types'
 import React, { Component } from 'react'
 import { useTreePath } from '../utils/treePath'
-// import ExtensionPoint from './ExtensionPoint'
-import ExtensionPoint from './ExtensionPointFC'
+import ExtensionPoint from './ExtensionPoint'
 import { useRuntime } from './RenderContext'
 
 type Element = string | ElementArray
