+++ react/core/main.tsx
@@ -19,8 +19,7 @@ import Loading from '../components/Loading'
 
 import { ChildBlock, useChildBlock } from '../components/ChildBlock'
 import ExtensionContainer from '../components/ExtensionContainer'
-// import ExtensionPoint from '../components/ExtensionPoint'
-import ExtensionPoint from '../components/ExtensionPointFC'
+import ExtensionPoint from '../components/ExtensionPoint'
 import LayoutContainer from '../components/LayoutContainer'
 import LegacyExtensionContainer from '../components/LegacyExtensionContainer'
 import Link from '../components/Link'
