+++ react/utils/dom.tsx
@@ -79,6 +79,8 @@ export const getMarkups = (
   return markups
 }
 
+export const isSiteEditorIframe = canUseDOM && window.top !== window.self && window.top.__provideRuntime && !!window.top.__provideRuntime
+
 export const getContainer = () => {
   return canUseDOM
     ? document.getElementsByClassName(RENDER_CONTAINER_CLASS)[0]
